+++ src/Exception/SidRuntimeException.php
@@ -1,6 +1,0 @@
-<?php
-namespace HalimonAlexander\Sid\Exception;
-
-class SidRuntimeException extends \RuntimeException{
-  
-}