+++ ChangeLog.md
@@ -1,5 +1,10 @@
 Changelog
 =========
+* 1.2 (2018-04-14)
+ * vocabulary will be used as callback function now
+ * added SidItemNotFound and VocabularyCallbackNotSet exceptions
+ * removed SidRuntimeException
+
 * 1.1 (2017-11-27)
  * isHidden is checked using regex from now
  * convertation from hidden to common using regex too